requirements.txt
requests
langchain
llama-index
streamlit
pandas
langchain_community
fastapi == 0.115.0
uvicorn
langserve
python-dotenv
google-generativeai
langchain-google-genai
sse_starlette
langchain_groq
gpt4all
faiss-cpu
langchain_huggingface
arxiv